Simple Document Reviewer/requirements.txt
langchain==0.3.20
langchain-community
pdfplumber
python-docx
openai
langchain-anthropic
python-dotenv==1.0.0 